Refactor context helpers into support file
